Listed products in home page

+++ aws_project/lib/models/productModel.dart
@@ -3,7 +3,7 @@ import 'dart:ffi';
 class ProductModel {
   final int id;
   final String nome;
-  final Double preco;
+  final Float preco;
 
   const ProductModel(this.id, this.nome, this.preco);
 }
